+++ x/metoken/keeper/intest/grpc_query_test.go
@@ -1,18 +1,19 @@
-package keeper_test
+package intest
 
 import (
 	"testing"
 
-	"github.com/umee-network/umee/v5/x/metoken/mocks"
+	"gotest.tools/v3/assert"
 
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/umee-network/umee/v5/x/metoken"
-	"gotest.tools/v3/assert"
+	"github.com/umee-network/umee/v5/x/metoken/mocks"
 )
 
 func TestQuerier_Params(t *testing.T) {
-	s := initKeeperTestSuite(t, nil, nil)
+	s := initTestSuite(t, nil, nil)
 	querier, ctx := s.queryClient, s.ctx
 
 	resp, err := querier.Params(ctx, nil)
@@ -25,7 +26,7 @@ func TestQuerier_Indexes(t *testing.T) {
 	index1 := mocks.StableIndex(mocks.MeUSDDenom)
 	index2 := mocks.StableIndex("me/EUR")
 
-	s := initKeeperTestSuite(t, []metoken.Index{index1, index2}, nil)
+	s := initTestSuite(t, []metoken.Index{index1, index2}, nil)
 	querier, ctx := s.queryClient, s.ctx
 
 	tcs := []struct {
@@ -77,7 +78,7 @@ func TestQuerier_Balances(t *testing.T) {
 	balance1 := mocks.ValidUSDIndexBalances(mocks.MeUSDDenom)
 	balance2 := mocks.ValidUSDIndexBalances("me/EUR")
 
-	s := initKeeperTestSuite(t, nil, []metoken.IndexBalances{balance1, balance2})
+	s := initTestSuite(t, nil, []metoken.IndexBalances{balance1, balance2})
 	querier, ctx := s.queryClient, s.ctx
 
 	tcs := []struct {
@@ -135,7 +136,7 @@ func TestQuerier_SwapFee_meUSD(t *testing.T) {
 	index := mocks.StableIndex(mocks.MeUSDDenom)
 	balances := mocks.ValidUSDIndexBalances(mocks.MeUSDDenom)
 
-	s := initKeeperTestSuite(t, []metoken.Index{index}, []metoken.IndexBalances{balances})
+	s := initTestSuite(t, []metoken.Index{index}, []metoken.IndexBalances{balances})
 	querier, ctx := s.queryClient, s.ctx
 
 	// set prices
@@ -212,7 +213,7 @@ func TestQuerier_RedeemFee_meUSD(t *testing.T) {
 	index := mocks.StableIndex(mocks.MeUSDDenom)
 	balances := mocks.ValidUSDIndexBalances(mocks.MeUSDDenom)
 
-	s := initKeeperTestSuite(t, []metoken.Index{index}, []metoken.IndexBalances{balances})
+	s := initTestSuite(t, []metoken.Index{index}, []metoken.IndexBalances{balances})
 	querier, ctx, app := s.queryClient, s.ctx, s.app
 
 	// set prices
@@ -291,7 +292,7 @@ func TestQuerier_IndexPrice(t *testing.T) {
 	stableBalance := mocks.EmptyUSDIndexBalances(mocks.MeUSDDenom)
 	nonStableBalance := mocks.EmptyNonStableIndexBalances(mocks.MeNonStableDenom)
 
-	s := initKeeperTestSuite(
+	s := initTestSuite(
 		t,
 		[]metoken.Index{stableIndex, nonStableIndex},
 		[]metoken.IndexBalances{stableBalance, nonStableBalance},