+++ x/metoken/keeper/redeem.go
@@ -3,12 +3,13 @@ package keeper
 import (
 	"fmt"
 
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
 	sdkmath "cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+
+	"github.com/umee-network/umee/v5/util/coin"
 	"github.com/umee-network/umee/v5/x/metoken"
+	"github.com/umee-network/umee/v5/x/metoken/errors"
 )
 
 // redeemResponse wraps all the coins of a successful redemption
@@ -63,20 +64,16 @@ func (k Keeper) redeem(userAddr sdk.AccAddress, meToken sdk.Coin, assetDenom str
 		return redeemResponse{}, err
 	}
 
-	uTokenWithdrawn, err := k.withdrawFromLeverage(sdk.NewCoin(assetDenom, amountFromLeverage))
+	tokensWithdrawn, err := k.withdrawFromLeverage(sdk.NewCoin(assetDenom, amountFromLeverage))
 	if err != nil {
 		return redeemResponse{}, err
 	}
 
 	// if there is a difference between the desired to withdraw from x/leverage and the withdrawn,
 	// take it from x/metoken reserves
-	tokenWithdrawn, err := k.leverageKeeper.ExchangeUToken(*k.ctx, uTokenWithdrawn)
-	if err != nil {
-		return redeemResponse{}, err
-	}
 
-	if tokenWithdrawn.Amount.LT(amountFromLeverage) {
-		tokenDiff := amountFromLeverage.Sub(tokenWithdrawn.Amount)
+	if tokensWithdrawn.Amount.LT(amountFromLeverage) {
+		tokenDiff := amountFromLeverage.Sub(tokensWithdrawn.Amount)
 		amountFromReserves = amountFromReserves.Add(tokenDiff)
 		amountFromLeverage = amountFromLeverage.Sub(tokenDiff)
 	}
@@ -148,28 +145,35 @@ func (k Keeper) redeem(userAddr sdk.AccAddress, meToken sdk.Coin, assetDenom str
 }
 
 // withdrawFromLeverage before withdrawing from x/leverage check if it's possible to withdraw the desired amount
-// based on x/leverage module constrains. When the full amount is not available withdraw the max possible.
+// based on x/leverage module constraints. When the full amount is not available withdraw the max possible.
+// Returning args are:
+//   - tokensWithdrawn: the amount tokens withdrawn from x/leverage.
+//   - error
 func (k Keeper) withdrawFromLeverage(tokensToWithdraw sdk.Coin) (sdk.Coin, error) {
-	uTokensFromLeverage, err := k.leverageKeeper.ExchangeToken(*k.ctx, tokensToWithdraw)
+	uTokensFromLeverage, err := k.leverageKeeper.ToUToken(*k.ctx, tokensToWithdraw)
 	if err != nil {
-		return sdk.Coin{}, err
+		return sdk.Coin{}, errors.Wrap(err, true)
 	}
 
 	availableUTokensFromLeverage, err := k.leverageKeeper.ModuleMaxWithdraw(*k.ctx, uTokensFromLeverage)
 	if err != nil {
-		return sdk.Coin{}, err
+		return sdk.Coin{}, errors.Wrap(err, true)
 	}
 
-	uTokensToWithdraw := sdk.NewCoin(uTokensFromLeverage.Denom, availableUTokensFromLeverage)
-	if _, _, err = k.leverageKeeper.Withdraw(
+	if availableUTokensFromLeverage.IsZero() {
+		return coin.Zero(tokensToWithdraw.Denom), nil
+	}
+
+	tokensWithdrawn, recoverable, err := k.leverageKeeper.WithdrawToModule(
 		*k.ctx,
-		authtypes.NewModuleAddress(metoken.ModuleName),
-		uTokensToWithdraw,
-	); err != nil {
-		return sdk.Coin{}, err
+		metoken.ModuleName,
+		sdk.NewCoin(uTokensFromLeverage.Denom, sdk.MinInt(availableUTokensFromLeverage, uTokensFromLeverage.Amount)),
+	)
+	if err != nil {
+		return sdk.Coin{}, errors.Wrap(err, recoverable)
 	}
 
-	return uTokensToWithdraw, nil
+	return tokensWithdrawn, nil
 }
 
 // calculateRedeem returns the fee to be charged to the user,